Extract payments UI into components/helpers
Move payments-related UI and utilities out of MyPayments into reusable modules. Adds paymentsHelpers.js (status/configs, formatters, CSS, isOverdue), ContractTabs.jsx, PaymentRow.jsx and PaymentsBadges.jsx (Toast, Skeleton, SectionLabel). Updates MyPayments.jsx to import these components and removes the duplicated constants/styles and inline components to improve modularity and reuse.
