+++ src/pages/client/MyPayments.jsx
@@ -2,193 +2,10 @@ import { useState, useEffect, useCallback, useContext } from "react";
 import MainLayout from "../../components/layout/Layout";
 import { AuthContext } from "../../context/AuthProvider";
 import api from "../../api/axios";
-
-// ─── Constants ────────────────────────────────────────────────────────────────
-const PAY_STATUS_CFG = {
-  PENDING:  { bg:"rgba(201,184,122,0.1)",  color:"#c9b87a", border:"rgba(201,184,122,0.25)", label:"Pending"  },
-  PAID:     { bg:"rgba(126,184,164,0.1)",  color:"#2a6049", border:"rgba(126,184,164,0.25)", label:"Paid"     },
-  FAILED:   { bg:"rgba(212,133,90,0.1)",   color:"#8b4513", border:"rgba(212,133,90,0.25)",  label:"Failed"   },
-  OVERDUE:  { bg:"rgba(212,133,90,0.12)",  color:"#8b3a1c", border:"rgba(212,133,90,0.3)",   label:"Overdue"  },
-  REFUNDED: { bg:"rgba(160,153,126,0.1)",  color:"#6b6248", border:"rgba(160,153,126,0.22)", label:"Refunded" },
-};
-
-const CONTRACT_STATUS_CFG = {
-  ACTIVE:            { label:"Active",            dot:"#7eb8a4", pill:"rgba(126,184,164,0.13)", pillBorder:"rgba(126,184,164,0.28)", color:"#2a6049"  },
-  ENDED:             { label:"Ended",             dot:"#a0997e", pill:"rgba(160,153,126,0.1)",  pillBorder:"rgba(160,153,126,0.22)", color:"#6b6248"  },
-  CANCELLED:         { label:"Cancelled",         dot:"#d4855a", pill:"rgba(212,133,90,0.1)",   pillBorder:"rgba(212,133,90,0.25)",  color:"#8b4513"  },
-  PENDING_SIGNATURE: { label:"Pending Signature", dot:"#c9b87a", pill:"rgba(201,184,122,0.12)", pillBorder:"rgba(201,184,122,0.28)", color:"#9a7a30"  },
-};
-
-const TYPE_ICON = { RENT:"💳", DEPOSIT:"🔒", LATE_FEE:"🔴", REFUND:"↩️" };
-
-const fmtDate  = (d) => d ? new Date(d).toLocaleDateString("en-GB", { day:"2-digit", month:"short", year:"numeric" }) : "—";
-const fmtMoney = (v) => v != null ? `€${Number(v).toLocaleString("de-DE")}` : "—";
-
-function isOverdue(p) {
-  if (!p.due_date || p.status !== "PENDING") return false;
-  return new Date(p.due_date) < new Date();
-}
-
-// ─── Global CSS ───────────────────────────────────────────────────────────────
-const CSS = `
-  @import url('https://fonts.googleapis.com/css2?family=Cormorant+Garamond:ital,wght@0,400;0,600;0,700;1,400&family=DM+Sans:wght@300;400;500;600&display=swap');
-
-  .cp * { box-sizing: border-box; }
-  .cp { font-family: 'DM Sans', system-ui, sans-serif; background: #f2ede4; min-height: 100vh; }
-
-  .cp-card { transition: transform 0.22s ease, box-shadow 0.22s ease; }
-  .cp-card:hover { transform: translateY(-4px); box-shadow: 0 20px 44px rgba(20,16,10,0.12) !important; }
-
-  .cp-pay-row { transition: box-shadow 0.15s ease, background 0.15s ease; }
-  .cp-pay-row:hover { box-shadow: 0 4px 18px rgba(20,16,10,0.08) !important; background: #faf7f2 !important; }
-
-  .cp-btn { transition: all 0.17s ease; }
-  .cp-btn:hover { opacity: 0.85; transform: translateY(-1px); }
-
-  .cp-chip { transition: all 0.14s ease; }
-
-  @keyframes cp-fade-up  { from{opacity:0;transform:translateY(18px)} to{opacity:1;transform:translateY(0)} }
-  @keyframes cp-shimmer  { 0%{background-position:-800px 0} 100%{background-position:800px 0} }
-  @keyframes cp-pulse    { 0%,100%{opacity:.38} 50%{opacity:.82} }
-  @keyframes cp-spin     { to{transform:rotate(360deg)} }
-  @keyframes cp-toast    { from{opacity:0;transform:translateY(10px)} to{opacity:1;transform:translateY(0)} }
-  @keyframes cp-glow     { 0%,100%{opacity:0.07} 50%{opacity:0.14} }
-  @keyframes cp-card-in  { from{opacity:0;transform:translateY(16px)} to{opacity:1;transform:translateY(0)} }
-`;
-
-// ─── Toast ────────────────────────────────────────────────────────────────────
-function Toast({ msg, type = "success", onDone }) {
-  useEffect(() => { const t = setTimeout(onDone, 3000); return () => clearTimeout(t); }, [onDone]);
-  return (
-    <div style={{
-      position:"fixed", bottom:26, right:26, zIndex:9999,
-      background:"#1a1714", color:type==="error"?"#f09090":"#90c8a8",
-      padding:"11px 18px", borderRadius:12, fontSize:13,
-      boxShadow:"0 10px 36px rgba(0,0,0,0.32)",
-      border:`1px solid ${type==="error"?"rgba(240,128,128,0.15)":"rgba(144,200,168,0.15)"}`,
-      maxWidth:320, fontFamily:"'DM Sans',sans-serif",
-      animation:"cp-toast 0.2s ease", display:"flex", alignItems:"center", gap:8,
-    }}>
-      <span style={{fontSize:14}}>{type==="error"?"⚠️":"✅"}</span>{msg}
-    </div>
-  );
-}
-
-// ─── Skeleton ─────────────────────────────────────────────────────────────────
-function Skeleton({ count=4 }) {
-  return (
-    <div style={{ display:"flex", flexDirection:"column", gap:10 }}>
-      {Array.from({ length:count }).map((_, i) => (
-        <div key={i} style={{
-          background:"linear-gradient(90deg,#ede9df 25%,#e4ddd0 50%,#ede9df 75%)",
-          backgroundSize:"800px 100%", borderRadius:12, height:68,
-          animation:"cp-shimmer 1.6s ease-in-out infinite",
-        }}/>
-      ))}
-    </div>
-  );
-}
-
-// ─── Contract Selector Tabs ───────────────────────────────────────────────────
-function ContractTabs({ contracts, selectedId, onSelect }) {
-  if (contracts.length === 0) return null;
-  return (
-    <div style={{ display:"flex", gap:7, flexWrap:"wrap" }}>
-      <button onClick={() => onSelect(null)} className="cp-chip"
-        style={{
-          padding:"7px 16px", borderRadius:999, fontSize:12.5, fontWeight:600,
-          border:`1.5px solid ${selectedId===null?"#1a1714":"#e4ddd0"}`,
-          background:selectedId===null?"#1a1714":"transparent",
-          color:selectedId===null?"#f5f0e8":"#6b6248",
-          cursor:"pointer", fontFamily:"inherit",
-        }}>
-        All Contracts
-      </button>
-      {contracts.map(c => {
-        const s      = CONTRACT_STATUS_CFG[c.status] || CONTRACT_STATUS_CFG.ENDED;
-        const active = selectedId === c.id;
-        return (
-          <button key={c.id} onClick={() => onSelect(c.id)} className="cp-chip"
-            style={{
-              padding:"7px 14px", borderRadius:999, fontSize:12.5, fontWeight:600,
-              border:`1.5px solid ${active?"#1a1714":"#e4ddd0"}`,
-              background:active?"#1a1714":"transparent",
-              color:active?"#f5f0e8":"#6b6248",
-              cursor:"pointer", fontFamily:"inherit",
-              display:"flex", alignItems:"center", gap:7,
-            }}>
-            Contract #{c.id}
-            <span style={{
-              background:active?"rgba(245,240,232,0.12)":s.pill,
-              color:active?"rgba(245,240,232,0.7)":s.color,
-              border:`1px solid ${active?"rgba(245,240,232,0.2)":s.pillBorder}`,
-              padding:"1px 8px", borderRadius:999, fontSize:10.5, fontWeight:700,
-            }}>
-              {s.label}
-            </span>
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
-// ─── Section Label ────────────────────────────────────────────────────────────
-function SectionLabel({ label, count, color }) {
-  return (
-    <div style={{ display:"flex", alignItems:"center", gap:8, marginBottom:8 }}>
-      <span style={{ fontSize:9.5, fontWeight:700, color, textTransform:"uppercase", letterSpacing:"1px" }}>{label}</span>
-      <span style={{ background:`${color}14`, color, border:`1px solid ${color}30`, padding:"1px 9px", borderRadius:999, fontSize:10.5, fontWeight:700 }}>{count}</span>
-      <div style={{ flex:1, height:"1px", background:`${color}18` }}/>
-    </div>
-  );
-}
-
-// ─── Payment Row ──────────────────────────────────────────────────────────────
-function PaymentRow({ payment, idx }) {
-  const overdue       = isOverdue(payment);
-  const displayStatus = (overdue && payment.status==="PENDING") ? "OVERDUE" : payment.status;
-  const s             = PAY_STATUS_CFG[displayStatus] || PAY_STATUS_CFG.PENDING;
-  const icon          = payment.status==="PAID" ? "✓" : overdue ? "⚠" : (TYPE_ICON[payment.payment_type]||"💳");
-
-  return (
-    <div className="cp-pay-row"
-      style={{
-        display:"flex", alignItems:"center", justifyContent:"space-between",
-        padding:"13px 16px", background:"#fff", borderRadius:12,
-        border:`1.5px solid ${overdue?"rgba(212,133,90,0.22)":"#ece6da"}`,
-        animation:`cp-card-in 0.3s ease ${Math.min(idx*0.04,0.3)}s both`,
-      }}>
-
-      {/* Icon */}
-      <div style={{ width:38, height:38, borderRadius:10, flexShrink:0, display:"flex", alignItems:"center", justifyContent:"center", background:s.bg, border:`1.5px solid ${s.border}`, fontSize:16, marginRight:14 }}>
-        {icon}
-      </div>
-
-      {/* Info */}
-      <div style={{ flex:1, minWidth:0 }}>
-        <div style={{ display:"flex", alignItems:"center", gap:7, marginBottom:4, flexWrap:"wrap" }}>
-          <span style={{ fontWeight:700, fontSize:15, color:"#1a1714", fontFamily:"'Cormorant Garamond',Georgia,serif" }}>{fmtMoney(payment.amount)}</span>
-          <span style={{ background:"#f0ece3", color:"#6b5f45", border:"1px solid #e0d8c8", padding:"2px 9px", borderRadius:999, fontSize:10.5, fontWeight:600, textTransform:"uppercase" }}>{payment.payment_type}</span>
-          <span style={{ background:"rgba(201,184,122,0.1)", color:"#c9b87a", border:"1px solid rgba(201,184,122,0.22)", padding:"2px 9px", borderRadius:999, fontSize:10.5, fontWeight:600 }}>#{payment.contract_id}</span>
-        </div>
-        <p style={{ fontSize:12, color:"#b0a890", margin:0 }}>
-          Due: {fmtDate(payment.due_date)}
-          {payment.paid_date && ` · Paid: ${fmtDate(payment.paid_date)}`}
-          {payment.payment_method && ` · ${payment.payment_method}`}
-          {payment.transaction_ref && ` · Ref: ${payment.transaction_ref}`}
-        </p>
-        {payment.notes && <p style={{ fontSize:11.5, color:"#b0a890", margin:"3px 0 0", fontStyle:"italic", fontFamily:"'Cormorant Garamond',Georgia,serif" }}>{payment.notes}</p>}
-      </div>
-
-      {/* Status badge */}
-      <div style={{ display:"flex", alignItems:"center", gap:7, flexShrink:0, marginLeft:12 }}>
-        {overdue && <span style={{ fontSize:11.5, color:"#d4855a", fontWeight:700 }}>Overdue</span>}
-        <span style={{ background:s.bg, color:s.color, border:`1.5px solid ${s.border}`, padding:"3px 12px", borderRadius:999, fontSize:10.5, fontWeight:700, textTransform:"uppercase" }}>{s.label}</span>
-      </div>
-    </div>
-  );
-}
+import { CSS, fmtMoney, isOverdue, CONTRACT_STATUS_CFG } from "../../components/client/payments/paymentsHelpers";
+import { Toast, Skeleton, SectionLabel } from "../../components/client/payments/PaymentsBadges";
+import ContractTabs from "../../components/client/payments/ContractTabs";
+import PaymentRow from "../../components/client/payments/PaymentRow";
 
 // ═══════════════════════════════════════════════════════════════════════════════
 // MAIN PAGE
